rename useServers
