+++ src/hooks/index.js
@@ -1,3 +1,3 @@
 export { default as useCallout } from './useCallout';
 export { default as useLocationReset } from './useLocationReset';
-export { default as useServers } from './useServers';
+export { default as useCentralServers } from './useCentralServers';
